[develop] working step wise simulation, but see --->
There's a small difference in outputs compared to using step_to
with 1 yr stepping:

yearly   step_to: 1.569239985, 7.069009343, 27.21039394
biweekly step_to: 1.569460052, 7.060768412, 27.17424297

This difference is due to insertion of an extra point in R0 moving
averager at integer t in yearly stepping - this doesnt happen at
biweekly stepping, which seems more correct. This is the best
level of investigattion I can do for now, so I'm going to take
this as the expected test result and proceed with other stuff.